Split examples to eliminate travis failures
