+++ Makefile
@@ -79,7 +79,7 @@ unit-tests:
 	@go test $(VERBOSE) $(UNIT_TEST_PACKAGES) -cover -coverprofile=cover.out
 
 .PHONY: e2e-tests
-e2e-tests: prepare-e2e-tests e2e-tests-smoke e2e-tests-cassandra e2e-tests-es e2e-tests-self-provisioned-es e2e-tests-streaming e2e-tests-examples
+e2e-tests: prepare-e2e-tests e2e-tests-smoke e2e-tests-cassandra e2e-tests-es e2e-tests-self-provisioned-es e2e-tests-streaming e2e-tests-examples1 e2e-tests-examples2
 
 .PHONY: prepare-e2e-tests
 prepare-e2e-tests: crd build docker push
@@ -120,10 +120,15 @@ e2e-tests-streaming: prepare-e2e-tests es kafka
 	@echo Running Streaming end-to-end tests...
 	@STORAGE_NAMESPACE=$(STORAGE_NAMESPACE) KAFKA_NAMESPACE=$(KAFKA_NAMESPACE) go test -tags=streaming ./test/e2e/... $(TEST_OPTIONS)
 
-.PHONY: e2e-tests-examples
-e2e-tests-examples: prepare-e2e-tests es cassandra kafka deploy-es-operator
+.PHONY: e2e-tests-examples1
+e2e-tests-examples1: prepare-e2e-tests es cassandra deploy-es-operator
 	@echo Running Example end-to-end tests...
-	@STORAGE_NAMESPACE=$(STORAGE_NAMESPACE) KAFKA_NAMESPACE=$(KAFKA_NAMESPACE) go test -tags=examples ./test/e2e/... $(TEST_OPTIONS)
+	@STORAGE_NAMESPACE=$(STORAGE_NAMESPACE) KAFKA_NAMESPACE=$(KAFKA_NAMESPACE) go test -tags=examples1 ./test/e2e/... $(TEST_OPTIONS)
+
+.PHONY: e2e-tests-examples2
+e2e-tests-examples2: prepare-e2e-tests kafka es
+	@echo Running Example end-to-end tests...
+	@STORAGE_NAMESPACE=$(STORAGE_NAMESPACE) KAFKA_NAMESPACE=$(KAFKA_NAMESPACE) go test -tags=examples2 ./test/e2e/... $(TEST_OPTIONS)
 
 .PHONY: run
 run: crd
@@ -233,7 +238,3 @@ install-tools:
 
 .PHONY: install
 install: install-sdk install-tools vendor
-
-.PHONY: setup-minikube
-setup-minikube:
-	@./.travis/setupMinikube.sh
